Rename character-info.ts
