+++ src/data/character-info.ts
